Created the sections structure

+++ app/Http/Controllers/GraderController.php
@@ -6,6 +6,8 @@ use App\Course;
 use App\Grader;
 use App\GraderAccessCode;
 use App\Http\Requests\StoreGrader;
+use App\Http\Requests\UpdateGrader;
+use App\Section;
 use App\User;
 use App\Exceptions\Handler;
 use \Exception;
@@ -15,7 +17,50 @@ use Illuminate\Support\Facades\Gate;
 
 class GraderController extends Controller
 {
-    public function store(StoreGrader $request, Grader $grader, GraderAccessCode $graderAccessCode)
+
+    public function update(UpdateGrader $request, Course $course, User $user, Grader $grader)
+    {
+
+        $response['type'] = 'error';
+        $course = Course::find($request->course_id);
+        $authorized = Gate::inspect('update', [$grader, $course]);
+
+          if (!$authorized->allowed()) {
+              $response['message'] = $authorized->message();
+              return $response;
+          }
+        try {
+            $data = $request->validated();
+            $course_id = $request->course_id;
+            $grader_user_id = $user->id;
+            $section_ids = Course::find($course_id)->sections->pluck('id')->toArray();
+            DB::beginTransaction();
+            $grader->whereIn('section_id', $section_ids)
+                ->where('user_id', $grader_user_id )
+                ->delete();
+            foreach ($data['selected_sections'] as $section_id) {
+                $grader = new Grader();
+                $grader->user_id = $grader_user_id;
+                $grader->section_id = $section_id;
+                $grader->save();
+            }
+            $response['message'] = "The grader's sections have been updated.";
+            $response['type'] = 'success';
+            DB::commit();
+        } catch (Exception $e) {
+            DB::rollback();
+            $h = new Handler(app());
+            $h->report($e);
+            $response['message'] = "There was an error updating the sections.  Please try again by refreshing the page or contact us for assistance.";
+        }
+        return $response;
+
+    }
+
+    public function store(StoreGrader $request,
+                          Grader $grader,
+                          GraderAccessCode $graderAccessCode,
+                          Section $section)
     {
 
         $response['type'] = 'error';
@@ -27,20 +72,27 @@ class GraderController extends Controller
         }
 
         $data = $request->validated();
+        $grader_acccess_codes = $graderAccessCode->where('access_code', $data['access_code'])
+            ->get();
+        if ($grader_acccess_codes->isEmpty()) {
+            $response['message'] = "There are no sections associated with that access code.";
+            return $response;
+        }
+        $course_section_names = [];
         try {
-
-            $course_id = $graderAccessCode->where('access_code', $data['access_code'])
-                ->get()
-                ->first()
-                ->course_id;
-            $course_name = Course::find($course_id)->name;
             DB::beginTransaction();
-            $grader->user_id = $request->user()->id;
-            $grader->course_id = $course_id;
-            $grader->save();
+            foreach ($grader_acccess_codes as $grader_acccess_code) {
+                $section = Section::find($grader_acccess_code->section_id);
+                $course_section_names[] = $section->course->name . ' - ' . $section->name;
+                $grader = new Grader();
+                $grader->user_id = $request->user()->id;
+                $grader->section_id = $section->id;
+                $grader->save();
+            }
+            $course_section_names = implode(', ', $course_section_names);
             DB::table('grader_access_codes')->delete(['access_code' => $data['access_code']]);
             DB::commit();
-            $response['message'] = "You have been added as a grader to <strong>$course_name</strong>.";
+            $response['message'] = "You have been added as a grader to <strong>$course_section_names</strong>.";
             $response['type'] = 'success';
         } catch (Exception $e) {
             DB::rollback();
@@ -91,18 +143,17 @@ class GraderController extends Controller
             $response['message'] = $authorized->message();
             return $response;
         }
+        $course_section_ids = $course->sections->pluck('id')->toArray();
         try {
-            $grader = $Grader
-                ->where('user_id', $student_user->id)
-                ->where('course_id', $course->id)
-                ->first();
-            $grader->delete();
+            $Grader->where('user_id', $student_user->id)->
+            whereIn('section_id', $course_section_ids)
+                ->delete();
             $response['type'] = 'success';
             $response['message'] = 'The grader has been removed from your course.';
         } catch (Exception $e) {
             $h = new Handler(app());
             $h->report($e);
-            $response['message'] = "There was an error retrieving your graders.  Please try again by refreshing the page or contact us for assistance.";
+            $response['message'] = "There was an error removing your grader.  Please try again by refreshing the page or contact us for assistance.";
         }
         return $response;
